+++ src/@types/index.d.ts
@@ -5,12 +5,12 @@ import * as cinerino from '@cinerino/domain';
 
 declare global {
     namespace Express {
-        export interface IRequestProject { typeOf: cinerino.factory.organizationType.Project; id: string; }
+        export interface IRequestProject { typeOf: cinerino.factory.chevre.organizationType.Project; id: string; }
 
         /**
          * APIユーザー(Cognitから認可を受ける)
          */
-        export type IUser = cinerino.factory.clientUser.IClientUser;
+        export type IUser = cinerino.factory.chevre.clientUser.IClientUser;
 
         // export type IRequestAgent = cinerino.factory.person.IPerson
         //     | cinerino.factory.creativeWork.softwareApplication.webApplication.ICreativeWork;